fix: failure when build directory contains spaces
Fix an installation issue that occurs when a path in which node-gyp is
run contains spaces.

For a path like `my app`, it would error with this message:

```
clang: error: no such file or directory: 'app/node_modules/node-addon-api'
```

The "my" part before the space (and the rest of the path before it)
would be missing from the path in the error message.

Fixes https://github.com/nodejs/node-gyp/issues/65