Make sure that INCS_* vars are properly escaped
Before this change, there was a chance that gyp-next would generate
broken Make files. Specifically, if the path to the project that you’re
building contains spaces, then the INCS_* Make variables would look like
this:

	INCS_Debug := \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/include/node \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/src \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/deps/openssl/config \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/deps/openssl/openssl/include \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/deps/uv/include \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/deps/zlib \
		-I/home/jayman/Documents/Home/VC/Git/Partially mine/node/repo/out/Debug/addons_headers/deps/v8/include

This would eventually cause builds to fail because the spaces in the
paths weren’t properly escaped. This change makes sure that the paths
get properly escaped.